rename controlles to follow naming convenation
